WIP: Make this a brainfuck interpreter
This pull-request changes from using a series of closures to run
a simple hand-made program to instead being a general purpose
brainfuck interpreter.

This is a naive approach and some of the code is not yet perfect
which is why this is WIP, that said it works.

The "standard benchmark" is the mandelbrot display, in this
implementation on my host I see 35 seconds to draw.  Compare that
to the naive VM I have which is about 47 seconds.

So we win, but it's still slow.

This is part of #1, but before merging:

* Remove the duplication on collapsing identical inc/dec/ops
* Test on more code.
* Linter updates.
* Better documentation.